backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url
django-cors-headers
python-dotenv
